Carrot_project/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
dj-database-url==0.5.0
Django==2.2.14
django-rest-framework==0.1.0
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
numpy==1.18.5
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
torch==1.5.1+cpu
torchvision==0.6.1+cpu
whitenoise==5.2.0
